Punishment timestamps are u32

+++ src/config/punish.rs
@@ -1,2 +1,2 @@
-pub const NO_RESPONSE_PUNISH: u64 = 5;
-pub const MAX_PUNISH: u64 = 10;
+pub const NO_RESPONSE_PUNISH: u32 = 5;
+pub const MAX_PUNISH: u32 = 10;
